server/requirements.txt
# FastAPI framework
fastapi
uvicorn[standard]

# Database
sqlalchemy
alembic
psycopg2-binary

# Authentication
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# Environment and configuration
python-dotenv
pydantic-settings
email-validator

# AI Integration
openai

# Cloud SQL (Google Cloud)
cloud-sql-python-connector
pg8000

# Logging and monitoring
structlog

# Testing
pytest
pytest-asyncio
httpx